Add function for validating webhook signatures
This commit adds a new function `ConstructEvent` which is used to decode the contents of
an HTTP request body into a type `WebhookEvent` that can be used to handle individual
webhook events.

The function also performs validation of the webhook signature prior to decoding. This
is done by generating an HMAC256 hash of the request contents and checking that it
matches the value provided in the `Signature` header.

Signed-off-by: David Bond <davidsbond93@gmail.com>